SSNVariability

Nothing to see here!

Examples

Linear case, match between numerics and analytic

Index

SSNVariability.diagtozero!

SSNVariability.jacobian

SSNVariability.norm_sum_rows!

SSNVariability.random_covariance_matrix

SSNVariability.run_network_nonoise

SSNVariability.run_network_to_convergence

SSNVariability.diagtozero! — Method

diagtozero!(M::AbstractMatrix{T}) where T

Replaces the matrix diagonal with zeros

SSNVariability.jacobian — Method

    jacobian(u,rn::RecurrentNetwork)

Jacobian matrix of the network dynamics at point u

SSNVariability.norm_sum_rows! — Method

norm_sum_rows!(mat)

Rescales the matrix by row so that the sum of each row is 1.0

SSNVariability.random_covariance_matrix — Function

random_covariance_matrix(dims::Integer,diag_val::Real,k_dims::Integer=5)

Returns a random covariance matrix that is positive definite and has off-diagonal elements.

Arguments

d: dimensions

diag_val: scaling of the diagonal

k-dims: to regulate off-diagonal elements

SSNVariability.run_network_nonoise — Method

function runnetworknonoise(ntw::RecurrentNeuralNetwork{R},rstart::Vector{R}, tend::Real; verbose::Bool=false,stepsize=0.05) where R<:Real

Runs the network simulation without any external noise.

Arguments

ntw::RecurrentNeuralNetwork

r_start::Vector : initial conditions

t_end::Real: the maximum time considered

Optional arguments

verbose::Bool = false : ODE solver is verbose

stepsize::Real = 0.05 : stepsize for saving the solution (not solver stepsize)

Outputs

times::Vector{R} : the times at which the solution is saved

u_end::Matrix{R} : the final state

r_end::Matrix{R} : the final state as rate

SSNVariability.run_network_to_convergence — Method

runnetworktoconvergence(u0, rn::RecurrentNeuralNetwork ; tend=80. , veltol=1E-4) Runs the network as described in run_network_nonoise, but stops as soon as norm(v) / n < veltol where v is the velocity at time t. If this condition is not satisfied (no convergence to attractor), it runs until t_end and prints a warning.

Arguments

rn::RecurrentNeuralNetwork

r_start::Vector : initial conditions

t_end::Real: the maximum time considered

veltol::Real : the norm (divided by num. dimensions) for velocity at convergence

Outputs

u_end::Vector : the final state at convergence

r_end::Vector : the final state at convergence as rate